requirements.txt.txt
torch
transformers
google-generativeai
googlemaps
geopy
timezonefinder
pytz
requests
python-dotenv
gradio